Challenge Project: Responsive Club Website.
In this project, you’ll be building your own club group page that will dynamically respond as you adjust the size of your screen!

In this project, you’ll build a webpage for a local club. You can base it on real-life organization you belong to or make one up! You’ll get to choose everything about your page: the club name, the page layout and styling, any images that you want to use, and more!